+++ src/components/Footer/styled.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components"
 
 export const FooterContainer = styled.footer`
-  grid-column: max / max 2;
+  grid-column: LB / RB 1;
   background: lightsteelblue;
 `
