+++ src/components/Header/styled.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components"
 
 export const HeaderContainer = styled.header`
-  grid-column: max / max 2;
+  grid-column: LB / RB 1;
   background: lightsteelblue;
   * {
     margin: 0;
