+++ src/components/Layout/styled.tsx
@@ -3,5 +3,5 @@ import styled from "styled-components"
 export const AppContainer = styled.div`
   margin: 0px;
   display: grid;
-  grid-template-columns: [max]1fr [XL]40px [M]50px[S] minmax(600px, 1fr) [S] 50px[M] 40px[XL] 1fr [max];
+  grid-template-columns: [LB] 1fr [LXL] 1fr [LS] minmax(600px, 1fr) [RS] 1fr [RXL] 1fr [RB];
 `
