+++ src/components/Ui/Container/index.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components"
 
 export const Container = styled.section`
-  grid-column: M / M 2;
+  grid-column: LS / RS 1;
   padding: 200px 0;
   border: 1px solid lightgray;
 `
