+++ src/pages/index.tsx
@@ -5,8 +5,10 @@ import SEO from "../components/Seo"
 import CourseGridItem from "../components/CourseGridItem"
 import Locales from "../components/Locales"
 import { Container } from "../components/Ui/Container"
+
 const IndexPage = props => {
   const { pages } = props.data
+
   return (
     <Layout>
       <SEO title="Home" />
@@ -18,12 +20,11 @@ const IndexPage = props => {
           <Locales location={"index.hero.content"} />
         </p>
       </Container>
+
       <Container>
         <h2>
           <Locales location={"ui.allCourses"} />
         </h2>
-      </Container>
-      <Container>
         {pages.edges
           .filter(node => node.node.context.course !== null)
           .map(course => (
