+++ src/templates/Course/styled.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components"
+
+export const CourseContainer = styled.div`
+  border: 2px solid gray;
+  display: block;
+  grid-column: LXL / RXL 1;
+  background: lightsteelblue;
+`
